added support for crate < 0.55
the `upper()` scalar function is only available in crate 0.55
it is used to group by query type, e.g. SELECT, UPDATE, INSERT, ...

so if ctop is now connected crate < 0.55 if will omit the `upper()`
function but this means that e.g. `select` and `SELECT` are listed
separately in the jobs log